feat(d2g): add volume mount artifact type

+++ tools/d2g/d2g.go
@@ -7,7 +7,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 	"testing"
@@ -157,7 +156,7 @@ func ConvertScopes(dwScopes []string, dwPayload *dockerworker.DockerWorkerPayloa
 		}
 	}
 
-	sort.Strings(gwScopes)
+	slices.Sort(gwScopes)
 
 	return
 }
@@ -303,7 +302,7 @@ func artifacts(dwPayload *dockerworker.DockerWorkerPayload) []genericworker.Arti
 		names[i] = name
 		i++
 	}
-	sort.Strings(names)
+	slices.Sort(names)
 	for i, name := range names {
 		gwArt := new(genericworker.Artifact)
 		defaults.SetDefaults(gwArt)
@@ -311,8 +310,15 @@ func artifacts(dwPayload *dockerworker.DockerWorkerPayload) []genericworker.Arti
 		gwArt.Expires = dwPayload.Artifacts[name].Expires
 		gwArt.Name = name
 		ext := filepath.Ext(dwPayload.Artifacts[name].Path)
-		gwArt.Path = "artifact" + strconv.Itoa(i) + ext
-		gwArt.Type = dwPayload.Artifacts[name].Type
+		if dwPayload.Artifacts[name].Type == "volume" {
+			gwArt.Path = "volume" + strconv.Itoa(i) + ext
+			// Generic Worker treats Docker Worker
+			// volume artifacts as a directory artifact
+			gwArt.Type = "directory"
+		} else {
+			gwArt.Path = "artifact" + strconv.Itoa(i) + ext
+			gwArt.Type = dwPayload.Artifacts[name].Type
+		}
 		gwArt.Optional = true
 
 		gwArtifacts[i] = *gwArt
@@ -334,11 +340,9 @@ func artifacts(dwPayload *dockerworker.DockerWorkerPayload) []genericworker.Arti
 }
 
 func command(dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, gwArtifacts []genericworker.Artifact, gwWritableDirectoryCaches []genericworker.WritableDirectoryCache, config map[string]any) ([][]string, string, error) {
-	containerName := ""
-	if testing.Testing() {
-		containerName = "taskcontainer"
-	} else {
-		containerName = "taskcontainer_" + slugid.Nice()
+	containerName := "taskcontainer"
+	if !testing.Testing() {
+		containerName = fmt.Sprintf("%s_%s", containerName, slugid.Nice())
 	}
 
 	commands := []string{}
@@ -357,7 +361,9 @@ func command(dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, gwArtif
 		)
 	}
 
-	runString, err := runCommand(containerName, dwPayload, dwImage, gwWritableDirectoryCaches, config)
+	copyArtifactCommands, captureExitCode := copyArtifacts(containerName, dwPayload, gwArtifacts)
+
+	runString, err := runCommand(containerName, dwPayload, dwImage, gwWritableDirectoryCaches, gwArtifacts, captureExitCode, config)
 	if err != nil {
 		return nil, containerName, fmt.Errorf("could not form docker run command: %w", err)
 	}
@@ -365,13 +371,20 @@ func command(dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, gwArtif
 	commands = append(
 		commands,
 		runString,
-		"exit_code=$?",
 	)
+
+	if captureExitCode {
+		commands = append(
+			commands,
+			"exit_code=$?",
+		)
+	}
 
 	commands = append(
 		commands,
-		copyArtifacts(containerName, dwPayload, gwArtifacts)...,
+		copyArtifactCommands...,
 	)
+
 	if dwPayload.Features.DockerSave {
 		commands = append(
 			commands,
@@ -380,11 +393,13 @@ func command(dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, gwArtif
 		)
 	}
 
-	commands = append(
-		commands,
-		"docker rm -v "+containerName,
-		`exit "${exit_code}"`,
-	)
+	if captureExitCode {
+		commands = append(
+			commands,
+			"docker rm -v "+containerName,
+			`exit "${exit_code}"`,
+		)
+	}
 
 	return [][]string{
 		{
@@ -395,11 +410,15 @@ func command(dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, gwArtif
 	}, containerName, nil
 }
 
-func runCommand(containerName string, dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, wdcs []genericworker.WritableDirectoryCache, config map[string]any) (string, error) {
+func runCommand(containerName string, dwPayload *dockerworker.DockerWorkerPayload, dwImage Image, wdcs []genericworker.WritableDirectoryCache, gwArtifacts []genericworker.Artifact, captureExitCode bool, config map[string]any) (string, error) {
 	command := strings.Builder{}
+	rm := " --rm"
+	if captureExitCode {
+		rm = ""
+	}
 	// Docker Worker used to attach a pseudo tty, see:
 	// https://github.com/taskcluster/taskcluster/blob/6b99f0ef71d9d8628c50adc17424167647a1c533/workers/docker-worker/src/task.js#L384
-	command.WriteString(fmt.Sprintf("timeout -s KILL %v docker run -t --name %v", dwPayload.MaxRunTime, containerName))
+	command.WriteString(fmt.Sprintf("timeout -s KILL %v docker run%s -t --name %v", dwPayload.MaxRunTime, rm, containerName))
 
 	// Do not limit resource usage by the containerName. See
 	// https://docs.docker.com/reference/cli/docker/container/run/
@@ -412,7 +431,7 @@ func runCommand(containerName string, dwPayload *dockerworker.DockerWorkerPayloa
 	if dwPayload.Capabilities.DisableSeccomp && config["allowDisableSeccomp"].(bool) {
 		command.WriteString(" --security-opt=seccomp=unconfined")
 	}
-	command.WriteString(createVolumeMountsString(dwPayload, wdcs, config))
+	command.WriteString(createVolumeMountsString(dwPayload, wdcs, gwArtifacts, config))
 	if dwPayload.Features.TaskclusterProxy && config["allowTaskclusterProxy"].(bool) {
 		command.WriteString(" --add-host=taskcluster:host-gateway")
 	}
@@ -426,8 +445,9 @@ func runCommand(containerName string, dwPayload *dockerworker.DockerWorkerPayloa
 	return command.String(), nil
 }
 
-func copyArtifacts(containerName string, dwPayload *dockerworker.DockerWorkerPayload, gwArtifacts []genericworker.Artifact) []string {
+func copyArtifacts(containerName string, dwPayload *dockerworker.DockerWorkerPayload, gwArtifacts []genericworker.Artifact) ([]string, bool) {
 	commands := []string{}
+	captureExitCode := dwPayload.Features.DockerSave
 	for i := range gwArtifacts {
 		// An image artifact will be in the generic worker payload when
 		// dockerSave is enabled. That artifact will not be found in either the
@@ -437,9 +457,14 @@ func copyArtifacts(containerName string, dwPayload *dockerworker.DockerWorkerPay
 		if _, ok := dwPayload.Artifacts[gwArtifacts[i].Name]; !ok {
 			continue
 		}
+		// Volume artifact mounts do not need to be copied
+		if dwPayload.Artifacts[gwArtifacts[i].Name].Type == "volume" {
+			continue
+		}
+		captureExitCode = true
 		commands = append(commands, fmt.Sprintf("docker cp %s:%s %s", containerName, shell.Escape(dwPayload.Artifacts[gwArtifacts[i].Name].Path), shell.Escape(gwArtifacts[i].Path)))
 	}
-	return commands
+	return commands, captureExitCode
 }
 
 func setEnv(dwPayload *dockerworker.DockerWorkerPayload, gwPayload *genericworker.GenericWorkerPayload) {
@@ -561,10 +586,15 @@ func fileNameWithoutExtension(fileName string) string {
 	return strings.TrimSuffix(fileName, filepath.Ext(fileName))
 }
 
-func createVolumeMountsString(dwPayload *dockerworker.DockerWorkerPayload, wdcs []genericworker.WritableDirectoryCache, config map[string]any) string {
+func createVolumeMountsString(dwPayload *dockerworker.DockerWorkerPayload, wdcs []genericworker.WritableDirectoryCache, gwArtifacts []genericworker.Artifact, config map[string]any) string {
 	volumeMounts := strings.Builder{}
 	for _, wdc := range wdcs {
 		volumeMounts.WriteString(` -v "$(pwd)/` + wdc.Directory + ":" + dwPayload.Cache[wdc.CacheName] + `"`)
+	}
+	for _, gwArtifact := range gwArtifacts {
+		if strings.HasPrefix(gwArtifact.Path, "volume") {
+			volumeMounts.WriteString(` -v "$(pwd)/` + gwArtifact.Path + ":" + dwPayload.Artifacts[gwArtifact.Name].Path + `"`)
+		}
 	}
 	if dwPayload.Capabilities.Devices.KVM && config["allowKVM"].(bool) {
 		volumeMounts.WriteString(" --device=/dev/kvm")
@@ -646,7 +676,7 @@ func envMappings(dwPayload *dockerworker.DockerWorkerPayload, config map[string]
 		envVarNames = append(envVarNames, envVarName)
 	}
 	envVarNames = append(envVarNames, additionalEnvVars...)
-	sort.Strings(envVarNames)
+	slices.Sort(envVarNames)
 	for _, envVarName := range envVarNames {
 		envStrBuilder.WriteString(envSetting(envVarName))
 	}
